+++ tvb/adapters/visualizers/brain.py
@@ -35,20 +35,15 @@
 """
 
 import os
-import json
 import numpy
 from tvb.adapters.visualizers.eeg_monitor import EegMonitor
-from tvb.adapters.visualizers.surface_view import ensure_shell_surface, SurfaceURLGenerator
+from tvb.adapters.visualizers.surface_view import ensure_shell_surface, SurfaceURLGenerator, ABCSurfaceDisplayer
 from tvb.adapters.visualizers.sensors import prepare_sensors_as_measure_points_params
 from tvb.adapters.visualizers.sensors import prepare_mapped_sensors_as_measure_points_params
 from tvb.core.entities.filters.chain import FilterChain
 from tvb.core.adapters.abcadapter import ABCAdapterForm
-from tvb.core.entities.model.datatypes.region_mapping import RegionMappingIndex
-from tvb.core.entities.model.datatypes.surface import SurfaceIndex
-from tvb.core.entities.model.datatypes.time_series import TimeSeriesIndex, TimeSeriesSurfaceIndex, TimeSeriesSEEGIndex, \
-    TimeSeriesRegionIndex, TimeSeriesEEGIndex
+from tvb.core.entities.model.datatypes.time_series import *
 from tvb.core.entities.storage import dao
-from tvb.core.adapters.abcdisplayer import ABCDisplayer
 from tvb.datatypes.surfaces import CORTICAL, EEG_CAP
 from tvb.core.neotraits._forms import DataTypeSelectField
 from tvb.core.neocom.h5 import DirLoader
@@ -81,7 +76,7 @@ class BrainViewerForm(ABCAdapterForm):
             operations=["in", "=="], values=[['TimeSeriesRegion', 'TimeSeriesSurface'], True])
 
 
-class BrainViewer(ABCDisplayer):
+class BrainViewer(ABCSurfaceDisplayer):
     """
     Interface between the 3D view of the Brain Cortical Surface and TVB framework.
     This viewer will build the required parameter dictionary that will be sent to the HTML / JS for further processing,
