+++ tvb/adapters/visualizers/pca.py
@@ -35,8 +35,8 @@ A displayer for the principal components analysis.
 
 """
 import json
+from tvb.adapters.visualizers.time_series import ABCSpaceDisplayer
 from tvb.core.adapters.abcadapter import ABCAdapterForm
-from tvb.core.adapters.abcdisplayer import ABCDisplayer
 from tvb.core.entities.model.datatypes.mode_decompositions import PrincipalComponentsIndex
 from tvb.core.neotraits._forms import DataTypeSelectField
 
@@ -61,7 +61,7 @@ class PCAForm(ABCAdapterForm):
         return PrincipalComponentsIndex
 
 
-class PCA(ABCDisplayer):
+class PCA(ABCSpaceDisplayer):
     _ui_name = "Principal Components Analysis Visualizer"
 
     def get_form_class(self):
